+++ R/createQTLMPP.R
@@ -94,6 +94,8 @@ print.summary.QTLMPP <- function(x,
 #' \item{QTLRegion}{A plot highlighting the QTLs found on the genetic map.}
 #' \item{QTLProfileExt}{A combination of the QTL profile and parental effects
 #' plotted above each other.}
+#' \item{parCIs}{A plot of the effect sizes and confidence intervals for each
+#' parent for each of the QTLs found.}
 #' }
 #' See details for a detailed description of the plots and the plot options
 #' specific to the different plots.
@@ -131,7 +133,7 @@ print.summary.QTLMPP <- function(x,
 #' \code{"Chromosomes"}}
 #' \item{\code{yLab}}{A character string, the y-axis label. Default =
 #' \code{"Parents"}}
-#' #' \item{\code{chr}}{A vector of chromosomes to be plotted. By default, all
+#' \item{\code{chr}}{A vector of chromosomes to be plotted. By default, all
 #' chromosomes are plotted. Using this option allows restricting the plot to a
 #' subset of chromosomes.}
 #' }
@@ -151,6 +153,12 @@ print.summary.QTLMPP <- function(x,
 #' chromosomes are plotted. Using this option allows restricting the plot to a
 #' subset of chromosomes.}
 #' }
+#'
+#' @section Parental Confidence Interval Plot:
+#' A plot of the parental effects and the confidence intervals around them for
+#' each of the QTLs found. Positive significant effects are shown in red,
+#' negative significant effects in blue, and all other effects are shown in
+#' black.
 #'
 #' @param x An object of class \code{QTLMPP}.
 #' @param ... further arguments to be passed on to the actual plotting
@@ -196,13 +204,16 @@ print.summary.QTLMPP <- function(x,
 #'
 #' ## Extended QTL Profile plot.
 #' plot(ABC_MQM, plotType = "QTLProfileExt")
+#'
+#' ## Parental confidence interval plot.
+#' plot(ABC_MQM, plotType = "parCIs")
 #' }
 #'
 #' @export
 plot.QTLMPP <- function(x,
                         ...,
                         plotType = c("QTLProfile", "parEffs", "QTLRegion",
-                                     "QTLProfileExt"),
+                                     "QTLProfileExt", "parCIs"),
                         title = NULL,
                         output = TRUE) {
   plotType <- match.arg(plotType)
@@ -319,6 +330,31 @@ plot.QTLMPP <- function(x,
       g1$widths[-1] <- maxWidths
       g2$widths[-1] <- maxWidths
       p <- gridExtra::grid.arrange(g1, g2)
+    } else if (plotType == "parCIs") {
+      parents <- x$GWASInfo$parents
+      trait <- GWAResult[["trait"]][1]
+      signSnp <- signSnp[signSnp[["snpStatus"]] == "significant SNP", ]
+      ## Create QTL effect data - basically a long format version of signSnp
+      parCIsDat <- lapply(X = parents, FUN = function(parent) {
+        parDat <- signSnp
+        parDat[["snp"]] <- paste("QTL at chr", parDat[["chr"]],
+                                 round(parDat[["pos"]], 0), "cM")
+        parDat[["parent"]] <- parent
+        parDat[["effect"]] <- parDat[[paste0("eff_", parent)]]
+        parDat[["seEffect"]] <- parDat[[paste0("se_eff_", parent)]]
+        return(parDat)
+      })
+      parCIsDat <- do.call(rbind, parCIsDat)
+      parCIsDat <- parCIsDat[c("snp", "parent", "effect", "seEffect")]
+      parCIsDat[["parent"]] <- factor(parCIsDat[["parent"]],
+                                         levels = parents)
+      p <- do.call(parCIsPlot,
+                   args = c(list(parCIsDat = parCIsDat,
+                                 title = title,
+                                 trait = trait,
+                                 output = output),
+                            dotArgs[!(names(dotArgs) == "parCIsDat")]))
+
     }
   }
   invisible(p)